+++ src/index.js
@@ -13,7 +13,6 @@ ReactDOM.render(
     
       <App />
     
-
     , document.getElementById('root'));
 
 // If you want your app to work offline and load faster, you can change
