+++ src/pages/users/UserDashboard.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import Profile from '../Profile'
+import Profile from '../../components/Profile'
 
 function UserDashboard() {
   return (
